feat(cdk/tree): move new demos to cdk-tree-redesign dir
