+++ src/components-examples/cdk/tree/cdk-tree-redesign/cdk-tree-flat-level-accessor-example.html
